Added jdoc renamed Datastore to DatastoreAccess
