+++ bootstrapper/src/main/java/ch/adv/ui/bootstrapper/Bootstrapper.java
@@ -23,6 +23,11 @@ public class Bootstrapper {
     private static final Logger logger = LoggerFactory.getLogger(Bootstrapper
             .class);
 
+    /**
+     * ADV UI entry point
+     *
+     * @param args cli arguments
+     */
     public static void main(String[] args) {
         logger.info("Bootstrapping ADV UI");
 
@@ -30,6 +35,10 @@ public class Bootstrapper {
         Application.launch(ADVApplication.class, args);
     }
 
+
+    /**
+     * Static map of all available modules
+     */
     static {
         Map<String, ADVModule> modules = new HashMap<>();
         modules.put("array", new ArrayModule());
