+++ ui-core/src/main/java/ch/adv/ui/ADVModule.java
@@ -1,6 +1,6 @@
 package ch.adv.ui;
 
-import ch.adv.ui.access.Datastore;
+import ch.adv.ui.access.DatastoreAccess;
 import ch.adv.ui.access.Parser;
 import ch.adv.ui.access.Stringifyer;
 import ch.adv.ui.presentation.Layouter;
@@ -13,5 +13,5 @@ public interface ADVModule {
     Layouter getLayouter();
     Stringifyer getStringifyer();
     Parser getParser();
-    Datastore getDatastore();
+    DatastoreAccess getDatastore();
 }
