+++ ui-core/src/main/java/ch/adv/ui/access/DatastoreAccess.java
@@ -2,7 +2,7 @@ package ch.adv.ui.access;
 
 import java.io.File;
 
-public interface Datastore {
+public interface DatastoreAccess {
 
     String read(File file);
     boolean write(File file, String json);}
